--- src/test/java/florizz/storage/LoadDBTest.java
@@ -1,4 +1,0 @@
-package florizz.storage;
-
-public class LoadDBTest {
-}
